Refactor : 폴더 위치 변경으로 import 수정, 더이상 사용하지 않는 app 스케줄러 삭제
