+++ closet-application/src/main/java/project/closet/service/waether/basic/BasicWeatherService.java
@@ -61,70 +61,6 @@ public class BasicWeatherService implements WeatherService {
         );
     }
 
-    @Override
-    @Scheduled(cron = "0 55 23 * * *", zone = "Asia/Seoul")
-    public void fetchAndSaveWeatherForecast() {
-        LocalDate forecastBaseDate = LocalDate.now();
-        LocalTime forecastTime = LocalTime.of(23, 0);
-        Instant forecastedAt = LocalDateTime.of(forecastBaseDate, forecastTime)
-            .atZone(SEOUL)
-            .toInstant();
-
-        fetchAndSave(forecastBaseDate, forecastTime, forecastedAt);
-    }
-
-    @Override
-    public void fetchAndSave(LocalDate baseDate, LocalTime forecastTime, Instant forecastedAt) {
-        log.info("🌤️ 날씨 정보 처리 요청: baseDate={}, forecastTime={}", baseDate, forecastTime);
-
-        List<WeatherLocation> locations = weatherLocationRepository.findAll();
-        int batchSize = 100;  // 원하는 batch 크기
-
-        for (int i = 0; i < locations.size(); i += batchSize) {
-            List<WeatherLocation> batch = locations.subList(i, Math.min(i + batchSize, locations.size()));
-            log.info("🚀 {}~{}번째 지역 날씨 요청 시작", i + 1, Math.min(i + batchSize, locations.size()));
-
-            List<CompletableFuture<List<Weather>>> futures = batch.stream()
-                .map(
-                    location -> weatherAPIClient.fetchWeatherAsync(
-                            location.getX(), location.getY(), baseDate, forecastTime
-                        )
-                        .thenApply(WeatherApiResponseConverter::parseToWeatherDataList)
-                        .thenApply(response ->
-                            weatherDataParser.parseToWeatherEntities(response, forecastedAt)
-                        )
-                        .exceptionally(ex -> {
-                            log.warn("❌ 날씨 요청 실패 (x={}, y={}): {}", location.getX(), location.getY(),
-                                ex.getMessage());
-                            return Collections.emptyList();
-                        })
-                ).toList();
-
-            // 이 batch가 완료될 때까지 기다림
-            List<Weather> parsedWeather = futures.stream()
-                .map(CompletableFuture::join)
-                .flatMap(List::stream)
-                .collect(Collectors.toList());
-
-            if (!parsedWeather.isEmpty()) {
-                weatherRepository.saveAll(parsedWeather);
-                log.info("✅ 배치 {}건 저장 완료", parsedWeather.size());
-            } else {
-                log.warn("⚠️ 저장할 데이터 없음 ({}~{})", i + 1, i + batchSize);
-            }
-
-            // 선택: 서버 과부하 방지용 sleep (필요시)
-            try {
-                Thread.sleep(1000);  // 1초 쉬었다가 다음 배치 실행
-            } catch (InterruptedException e) {
-                Thread.currentThread().interrupt();
-            }
-        }
-
-        log.info("🌤️ 날씨 정보 전체 처리 완료");
-    }
-
-
     // TODO 바로 전 날짜의 온도와 비교해서 온도 정보 반환
     @Transactional(readOnly = true)
     @Override
